Use GLM instead of own implementation and fix crashes

+++ src/objects/hitables/bvh_node.hpp
@@ -7,7 +7,7 @@
 class bvh_node : public hitable{
 public:
 	bvh_node() {}
-	bvh_node(hitable** l, int n, float time0, float time1);
+	bvh_node(hitable** l, size_t n, float time0, float time1);
 	virtual bool hit(const ray& r, float tmin, float tmax, hit_record& rec) const;
 	virtual bool bounding_box(float t0, float t1, aabb& box) const;
 	hitable* left;
